+++ docs/CHANGELOG.md
@@ -2,16 +2,21 @@
 
 # v0.26.0: Using the new github.com/gomlx/go-xla library.
 
-API Change: `dtypes` library changed from `github.com/gomlx/gopjrt/dtypes` to `github.com/gomlx/go-xla/pkg/types/dtypes`.
+API Change: `dtypes` package moved from `github.com/gomlx/gopjrt/dtypes` to `github.com/gomlx/go-xla/pkg/types/dtypes`.
 It should be a simple change in import.
 
+XLA:
 - go-xla (replacing the now deprecated stablehlo and gopjrt libraries)
   - Added auto-installation of standard (CPU and GPU/TPU when available) plugins.
     (Can be disabled by setting the environment variable `GOMLX_NO_AUTO_INSTALL` to anything)
   - Fixed some memory leaks on plugin destruction; 
-  - Improved performance in some low-latency scenarios (using GenPool as opposed to sync.Pool)
-- Removed old `gomlx/backends/xla`, using the now retired `xlabuilder` API for XLA.
+  - Improved performance in some low-latency scenarios (using GenPool as opposed to sync.Pool):
+- Removed old `gomlx/backends/xla` (the one that used the retired `xlabuilder` API for XLA).
 - Renamed `gomlx/backends/stablehlo` --> `gomlx/backends/xla`, using the new `go-xla` library.
+- Added `xla.EnableAutoInstall(enabled bool)` to enable/disable auto-installation of standard plugins.
+  And added `xla.AutoInstall()` to immediately auto-install standard plugins.
+
+Other updates:
 - Package `tensors`:
   - Added `CopyFlatData()` that returns an error (it was previously renamed to `MustCopyFlatData`)